// src/canisters/model/dfusion.did.d.ts
export interface UserExt {
  'id' : Principal,
  'bio' : [] | [string],
  'favorites' : Array<bigint>,
  'name' : [] | [string],
  'likes' : Array<bigint>,
  'entries' : Array<bigint>,
  'followers' : Array<Principal>,
  'following' : Array<Principal>,
}